feat: extract ingredient endpoints
