+++ backend/src/recipes/recipe.controller.spec.ts
@@ -1,5 +1,4 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { IngredientEntity } from './ingredient.entity';
 import { RecipeController } from './recipe.controller';
 import { RecipeEntity } from './recipe.entity';
 import { RecipeService } from './recipe.service';
@@ -18,17 +17,6 @@ describe('RecipeController', () => {
     ingredients: [],
   };
 
-  const mockIngredient: IngredientEntity = {
-    id: 1,
-    recipeId: 1,
-    recipe: {} as RecipeEntity,
-    name: 'Milk',
-    amount: 0.5,
-    unit: 'Liters',
-    createdAt: new Date('2026-01-01'),
-    updatedAt: new Date('2026-01-01'),
-  };
-
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [RecipeController],
@@ -41,8 +29,6 @@ describe('RecipeController', () => {
             create: jest.fn().mockResolvedValue(mockRecipe),
             update: jest.fn().mockResolvedValue(mockRecipe),
             remove: jest.fn().mockResolvedValue(undefined),
-            getIngredients: jest.fn().mockResolvedValue([mockIngredient]),
-            addIngredients: jest.fn().mockResolvedValue([mockIngredient]),
           },
         },
       ],
@@ -85,29 +71,6 @@ describe('RecipeController', () => {
       const result = await controller.remove(1);
       expect(result.message).toContain('deleted');
       expect(service.remove).toHaveBeenCalledWith(1);
-    });
-  });
-
-  describe('getIngredients', () => {
-    it('should return ingredients for a recipe', async () => {
-      const result = await controller.getIngredients(1);
-      expect(result.data).toHaveLength(1);
-      expect(result.data[0].name).toBe('Milk');
-      expect(service.getIngredients).toHaveBeenCalledWith(1);
-    });
-  });
-
-  describe('addIngredients', () => {
-    it('should add ingredients and return the updated list', async () => {
-      const payload = [
-        { name: 'Milk', amount: 0.5, unit: 'Liters' },
-        { name: 'Eggs', amount: 1, unit: 'pieces' },
-      ];
-
-      const result = await controller.addIngredients(1, payload);
-
-      expect(result.data).toHaveLength(1);
-      expect(service.addIngredients).toHaveBeenCalledWith(1, payload);
     });
   });
 });
