+++ backend/src/recipes/recipe.controller.ts
@@ -1,42 +1,8 @@
-import { ApiResponse, Ingredient, Recipe } from '@app/shared';
-import {
-  Body,
-  Controller,
-  Delete,
-  Get,
-  Param,
-  ParseArrayPipe,
-  ParseIntPipe,
-  Post,
-  Put,
-} from '@nestjs/common';
-import { CreateIngredientDto, CreateRecipeDto, UpdateRecipeDto } from './dto';
-import { IngredientEntity } from './ingredient.entity';
-import { RecipeEntity } from './recipe.entity';
+import { ApiResponse, Recipe } from '@app/shared';
+import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Put } from '@nestjs/common';
+import { CreateRecipeDto, UpdateRecipeDto } from './dto';
 import { RecipeService } from './recipe.service';
-
-function toRecipe(entity: RecipeEntity): Recipe {
-  return {
-    id: entity.id,
-    title: entity.title,
-    description: entity.description,
-    author: entity.author,
-    createdAt: entity.createdAt.toISOString(),
-    updatedAt: entity.updatedAt.toISOString(),
-  };
-}
-
-function toIngredient(entity: IngredientEntity): Ingredient {
-  return {
-    id: entity.id,
-    recipeId: entity.recipeId,
-    name: entity.name,
-    amount: entity.amount,
-    unit: entity.unit,
-    createdAt: entity.createdAt.toISOString(),
-    updatedAt: entity.updatedAt.toISOString(),
-  };
-}
+import { toRecipe } from './recipes.mapper';
 
 @Controller('api/recipes')
 export class RecipeController {
@@ -74,26 +40,5 @@ export class RecipeController {
   async remove(@Param('id', ParseIntPipe) id: number): Promise<{ message: string }> {
     await this.recipeService.remove(id);
     return { message: `Todo #${id} deleted` };
-  }
-
-  @Get(':id/ingredients')
-  async getIngredients(@Param('id', ParseIntPipe) id: number): Promise<ApiResponse<Ingredient[]>> {
-    const ingredients = await this.recipeService.getIngredients(id);
-    return { data: ingredients.map(toIngredient) };
-  }
-
-  @Post(':id/ingredients')
-  async addIngredients(
-    @Param('id', ParseIntPipe) id: number,
-    @Body(
-      new ParseArrayPipe({
-        items: CreateIngredientDto,
-        whitelist: true,
-      }),
-    )
-    dtos: CreateIngredientDto[],
-  ): Promise<ApiResponse<Ingredient[]>> {
-    const ingredients = await this.recipeService.addIngredients(id, dtos);
-    return { data: ingredients.map(toIngredient) };
   }
 }
